Also move ValidationResults utils
To common utils directory.
